Fix RHEL 7.3 SELinux compatibility
In RHEL7.3 the SELinux userspace was rebased to version 2.5, this was a breaking change and required us to update our policy to be compatible.
The new SELinux system in RHEL7.3 translates the policy into an intermediate language called CIL, which has more restrictions than the old system.
In order to be compatible we need to ensure that no system users or roles are re-declared in our policy.
CentOS6 excluded because SELinux is unchanged in that release.