Update SELinux policy variant support and improve installation error reporting
Removed support for the mls and minimum variants. They don't contain either the
unconfined_t or httpd_t type so there is no way we can make things work with them.

Ensured that we don't install the Passenger SELinux policy file for any variants
that we don't support.

Ensured that we build Passenger SELinux policy files for a distro-specific static
list of variants. The autodetection code was buggy.